refactor code organization
