+++ go.mod
@@ -1,5 +1,15 @@
-module github.com/dlorenc/gobump
+module github.com/chainguard-dev/gobump
 
 go 1.21
 
-require golang.org/x/mod v0.14.0
+require (
+	github.com/spf13/cobra v1.8.0
+	golang.org/x/mod v0.14.0
+	sigs.k8s.io/release-utils v0.7.7
+)
+
+require (
+	github.com/common-nighthawk/go-figure v0.0.0-20210622060536-734e95fb86be // indirect
+	github.com/inconshreveable/mousetrap v1.1.0 // indirect
+	github.com/spf13/pflag v1.0.5 // indirect
+)
