+++ pkg/types/types.go
@@ -1,0 +1,8 @@
+package types
+
+type Package struct {
+	Name    string
+	Version string
+	Replace bool
+	Require bool
+}
